+++ chef/jenkins-master/CHANGELOG.md
@@ -6,6 +6,5 @@ This file is used to list changes made in each version of the jenkins-master coo
 
 Initial release.
 
-- change 0
-- change 1
+- Jenkins master behind nginx